Made improvements to the current movement.
I made improvements to the way the camera moves to follow the player
(much smoother now).
I made the player movement a lot smoother and added a ground check using
raycasting which disables the trail renderers  when the player is not on
the ground.
Also made a little ramp for testing purposes.